WGLUtil->WGLInterface
